requirements.txt
torch>=1.10.0
torchvision>=0.11.0
ftfy
regex
tqdm
gdown
gradio
Pillow
numpy
matplotlib
lpips
opencv-python
git+https://github.com/openai/CLIP.git
